+++ src/components/product-card/reviewList.tsx
@@ -2,28 +2,35 @@
 import React, { useState } from 'react';
 import { Formik, Field, Form, ErrorMessage } from 'formik';
 import * as Yup from 'yup';
-import Rating from '@mui/material/Rating';
 import Send from '../../../public/icons/send.svg';
-import ReviewIcon from './../../../public/icons/reviews-icon.svg';
+import ReviewItem from './reviewItem';
+import Rating from '@mui/material/Rating';
 import IconStar from './../../../public/icons/rating-icon.svg';
 import IconStarEmpty from './../../../public/icons/rating-empty-icon.svg';
 
 type Props = {};
-const text = `I absolutely adore my Funko Nick Wilde POP Disney figure! The
+
+const reviews = [
+  `I absolutely adore my Funko Nick Wilde POP Disney figure! The
 attention to detail is fantastic, capturing Nick's mischievous
 grin and personality perfectly. The colors are vibrant, and the
 size is just right for displaying on my shelf I absolutely adore
 my Funko Nick Wilde POP Disney figure! The attention to detail
 is fantastic, capturing Nick's mischievous grin and personality
 perfectly. The colors are vibrant, and the size is just right
-for displaying on my shelf`;
+for displaying on my shelf`,
+  `I absolutely adore my Funko Nick Wilde POP Disney figure! The
+attention to detail is fantastic, capturing Nick's mischievous
+grin and personality perfectly. The colors are vibrant, and the
+size is just right for displaying on my shelf I absolutely adore
+my Funko Nick Wilde POP Disney figure! The attention to detail
+is fantastic, capturing Nick's mischievous grin and personality
+perfectly. The colors are vibrant, and the size is just right
+for displaying on my shelf`,
+];
+
 const CardReviews = (props: Props) => {
   const [ratingValue, setRatingValue] = useState<number | null>(0);
-  const [isShowMore, setIsShowMore] = useState(false);
-
-  const toggleReadMoreLess = () => {
-    setIsShowMore(!isShowMore);
-  };
 
   return (
     <div className="flex flex-col gap-5 px-[16px] py-5 rounded-[5px] bg-[#F5F5F5]">
@@ -72,64 +79,11 @@ const CardReviews = (props: Props) => {
         </Form>
       </Formik>
       <ul className="flex flex-col gap-5  overflow-auto h-[292px]">
-        <li className="flex gap-4">
-          <div className="flex flex-shrink-0 justify-center items-center w-10 h-10 rounded-[50%] bg-[#D9D9D9]">
-            <ReviewIcon />{' '}
-          </div>
-          <div>
-            <div className="flex justify-between">
-              <p className='text-base font-semibold'>Alison</p>
-              <Rating
-                readOnly={true}
-                name="rating-read"
-                emptyIcon={<IconStarEmpty />}
-                icon={<IconStar />}
-                value={4}
-              />
-            </div>
-            <div className="w-[284px]">
-              {!isShowMore && (
-                <p className="text-xs font-medium">{text.slice(0, 230)}...</p>
-              )}
-              {isShowMore && <p className="text-xs font-medium">{text}</p>}
-              <button
-                onClick={toggleReadMoreLess}
-                className="underline text-xs font-semibold"
-              >
-                {isShowMore ? 'Read less' : 'Read more'}
-              </button>
-            </div>
-          </div>
-        </li>
-        <li className="flex gap-4">
-          <div className="flex flex-shrink-0 justify-center items-center w-10 h-10 rounded-[50%] bg-[#D9D9D9]">
-            <ReviewIcon />{' '}
-          </div>
-          <div>
-            <div className="flex justify-between">
-              <p className="text-base font-semibold">Chris</p>
-              <Rating
-                readOnly={true}
-                name="rating-read"
-                emptyIcon={<IconStarEmpty />}
-                icon={<IconStar />}
-                value={3}
-              />
-            </div>
-            <div className="w-[284px]">
-              {!isShowMore && (
-                <p className="text-xs font-medium">{text.slice(0, 230)}...</p>
-              )}
-              {isShowMore && <p className="text-xs font-medium">{text}</p>}
-              <button
-                onClick={toggleReadMoreLess}
-                className="underline text-xs font-semibold"
-              >
-                {isShowMore ? 'Read less' : 'Read more'}
-              </button>
-            </div>
-          </div>
-        </li>
+        {reviews.map((item, index) => (
+          <li key={index} className="flex gap-4">
+            <ReviewItem item={item} />
+          </li>
+        ))}
       </ul>
     </div>
   );
